+++ android/src/main/java/com/github/fujianlian/klinechart/KLineEntity.java
@@ -134,7 +134,7 @@ public class KLineEntity implements IKLine {
 
     public List<Map<String, Object>> selectedItemList = new ArrayList<>();
 
-    public float id;
+    public double id;
     public String Date;
     public float Open;
     public float High;
